I'm gonna do what's called a pro-gamer move
